+++ package.json
@@ -11,6 +11,7 @@
     "lint-staged": "^9.4.1",
     "node": "^12.11.0",
     "nodemon": "^1.19.3",
+    "objection": "^1.6.11",
     "pg": "^7.12.1",
     "prettier": "^1.18.2",
     "ts-jest": "^24.1.0",
